requirements.txt
opencv-python
pytesseract
Pillow
streamlit
numpy
spacy==3.5.3
python-dotenv==1.0.0
pandas
fpdf==1.7.2
scikit-image==0.21.0
transformers==4.30.2
torch
tensorflow==2.12.0
# Additional languages for spaCy (optional)
# python -m spacy download en_core_web_lg
# python -m spacy download fr_core_news_sm
# python -m spacy download de_core_news_sm
# python -m spacy download es_core_news_sm